raiden+/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
boto3
PyGithub
langchain-google-genai
google-generativeai
google-genai
langchain-together
langchain-groq
langchain-deepseek
langchain-community
langchain-huggingface
langchain
langgraph
wikipedia
tavily-python
python-multipart
langchain-docling
pypdf
python-docx
unstructured
markdown
sentence-transformers
chromadb
faiss-cpu
tiktoken
youtube_search
langchain-chroma
langchain-experimental
pyinstaller
pywebview
importlib-metadata
psutil
pystray
pillow
keyring
torch
torchvision
transformers
pysqlite3-binary
onnxruntime
optimum
requests
dnspython
speedtest-cli
pypdf2
reportlab
pyOpenWeatherMap